ghostdog: Move error module to its own file
This moves the error module to its own file to prepare for more modules.
